+++ apps/client/src/app/modules/tooltip/xiv-ui-text.pipe.ts
@@ -1,49 +1,34 @@
-import { OnDestroy, Pipe, PipeTransform } from '@angular/core';
-import { Subscription } from 'rxjs';
-import { UIColor } from './ui-color';
+import { Pipe, PipeTransform } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
-import { UiColorsService } from './ui-colors.service';
 
 @Pipe({
   name: 'xivUIText',
-  pure: false
+  pure: true
 })
-export class UiTextPipe implements PipeTransform, OnDestroy {
+export class UiTextPipe implements PipeTransform {
 
-  private subscription: Subscription;
-
-  private colors: UIColor[] = [];
-
-  constructor(private uiColorsService: UiColorsService, private sanitizer: DomSanitizer) {
-    this.subscription = this.uiColorsService.getColors().subscribe(colors => this.colors = colors);
+  constructor(private sanitizer: DomSanitizer) {
   }
 
   transform(value: any = ''): SafeHtml {
     value = value.replace(/<73>(.*?)<\/73>/gi, '')
       .replace(/\n\n/g, '\n')
-      .replace('<SoftHyphen/>', "\u00AD")
+      .replace('<SoftHyphen/>', '\u00AD')
       .replace(/<Clickable\((.*?)\)\/>/, '$1')
       .replace(/<Highlight>.*?<\/Highlight>/, '')
       .replace(/<Switch.*?><Case\(1\)>(.*?)<\/Case>.*?<\/Switch>/, '$1')
       .replace(/<If.*?>(.*?)<Else\/>.*?<\/If>(?!<\/If|<Else)/, '$1')
       .replace(/<\/?Emphasis>/, '*')
+      .replace(/<UIForeground>([^<\/]*)<\/UIForeground>/gi, '<span>')
+      .replace(/<UIGlow>([^<\/]*)<\/UIGlow>/gi, '<span>')
       .replace('<Indent/>', ' ')
       .replace(/<72>01<\/72>/gi, '</span>')
       .replace(/<UI\w+>01<\/UI\w+>/gi, '</span>')
       .replace(/\n/g, '<br>')
-      .replace(/\s{2,}/, ' ');
-    let match;
-    while (match = /<UI\w+>(.*?)<\/UI\w+>/gi.exec(value)) {
-      const colorID = parseInt(match[1].substr(-4), 16);
-      const colorEntry: UIColor = this.colors.find(c => c.ID === colorID);
-      const color = colorEntry === undefined ? 'fff' : colorEntry.ColorAHex;
-      value = value.replace(match[0], `<span style="color: #${color};">`);
-    }
+      .replace(/\s{2,}/, ' ')
+      .replace(/<span style="color:#0001f4;">([^<\/]*)<\/span>/gmi, '<b>$1</b>')
+      .replace(/<span style="color:#0001f8;">([^<\/]*)<\/span>/gmi, '<b>$1</b>');
     return this.sanitizer.bypassSecurityTrustHtml(value);
-  }
-
-  ngOnDestroy(): void {
-    this.subscription.unsubscribe();
   }
 
 }