+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="5.4.6"></a>
+## [5.4.6](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/compare/v5.4.5...v5.4.6) (2019-05-25)
+
+
+### Bug Fixes
+
+* **db:** fixed trait descriptions colors ([ca758a5](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/ca758a5))
+* **desktop:** fixed a possible crash when applying settings ([0144f87](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/0144f87))
+
+
+### Features
+
+* **alarms:** you can now see the note inside the alarms overlay ([4029497](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/4029497))
+* Teamcraft now displays random tips inside the topbar ([2563962](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/2563962))
+* **log-tracker:** added item name color based on rarity ([fe12751](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/fe12751))
+* **map:** added aetheryte shard names and more accurate aetheryte names overall ([a90549d](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/a90549d))
+
+
+
+<a name="5.4.5"></a>
+## [5.4.5](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/compare/v5.4.4...v5.4.5) (2019-05-24)
+
+
+### Bug Fixes
+
+* **list:** fixed an issue with progression when an item is final and material ([c4bef2b](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/c4bef2b))
+* **perf:** performance improvements for UI text pipe ([072488f](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/072488f))
+
+
+### Features
+
+* **db:** added text color for items based on rarity ([0a1c1c0](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/0a1c1c0))
+* **db:** item pages now include a model viewer from garlandtools ([1aa1701](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/1aa1701))
+* **db:** items that are obtained from reduction now have alarm integration ([c088579](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/commit/c088579))
+
+
+
 <a name="5.4.4"></a>
 ## [5.4.4](https://github.com/ffxiv-teamcraft/ffxiv-teamcraft/compare/v5.4.2...v5.4.4) (2019-05-21)